+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.1.1
+
+- 成功保留了id和一个空格
+- 修正部分文本
+- 现在可以适配标题和块引用
+- Keep id and one space successed
+- Correct some texts
+- Now fit title and block reference
+
 ## v0.1.0
 
 - 成功做到了将空格去掉，后续等待其他适配